feat: commons dialog
