+++ core/ui-common/src/main/kotlin/com/wire/android/ui/common/WireDialog.kt
@@ -31,17 +31,13 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.ClickableText
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.platform.LocalUriHandler
@@ -49,9 +45,7 @@ import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import com.wire.android.ui.common.button.WireButtonState
@@ -59,9 +53,6 @@ import com.wire.android.ui.common.button.WirePrimaryButton
 import com.wire.android.ui.common.button.WireSecondaryButton
 import com.wire.android.ui.common.button.WireTertiaryButton
 import com.wire.android.ui.common.progress.WireCircularProgressIndicator
-import com.wire.android.ui.common.textfield.WirePasswordTextField
-import com.wire.android.ui.markdown.MarkdownConstants
-import com.wire.android.ui.theme.WireTheme
 import com.wire.android.ui.theme.wireColorScheme
 import com.wire.android.ui.theme.wireDimensions
 import com.wire.android.ui.theme.wireTypography
@@ -160,7 +151,7 @@ fun WireDialog(
 }
 
 @Composable
-private fun WireDialogContent(
+fun WireDialogContent(
     title: String,
     titleLoading: Boolean = false,
     text: AnnotatedString? = null,
@@ -209,7 +200,7 @@ private fun WireDialogContent(
                             modifier = Modifier.padding(bottom = MaterialTheme.wireDimensions.dialogTextsSpacing),
                             onClick = { offset ->
                                 text.getStringAnnotations(
-                                    tag = MarkdownConstants.TAG_URL,
+                                    tag = TAG_URL,
                                     start = offset,
                                     end = offset,
                                 ).firstOrNull()?.let { result -> uriHandler.openUri(result.item) }
@@ -299,143 +290,6 @@ private fun WireDialogButtonProperties?.getButton(modifier: Modifier = Modifier)
     }
 }
 
-@OptIn(ExperimentalComposeUiApi::class)
-@Preview(showBackground = true)
-@Composable
-fun PreviewWireDialog() {
-    var password by remember { mutableStateOf(TextFieldValue("")) }
-    WireTheme {
-        Box(
-            contentAlignment = Alignment.Center,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            WireDialogContent(
-                optionButton1Properties = WireDialogButtonProperties(
-                    text = "OK",
-                    onClick = { },
-                    type = WireDialogButtonType.Primary,
-                    state = if (password.text.isEmpty()) WireButtonState.Disabled else WireButtonState.Error,
-                ),
-                dismissButtonProperties = WireDialogButtonProperties(
-                    text = "Cancel",
-                    onClick = { }
-                ),
-                title = "title",
-                text = buildAnnotatedString {
-                    val style = SpanStyle(
-                        color = colorsScheme().onBackground,
-                        fontWeight = MaterialTheme.wireTypography.body01.fontWeight,
-                        fontSize = MaterialTheme.wireTypography.body01.fontSize,
-                        fontFamily = MaterialTheme.wireTypography.body01.fontFamily,
-                        fontStyle = MaterialTheme.wireTypography.body01.fontStyle
-                    )
-                    withStyle(style) { append("text\nsecond line\nthirdLine\nfourth line\nfifth line\nsixth line\nseventh line") }
-                },
-            ) {
-                WirePasswordTextField(
-                    value = password,
-                    onValueChange = { password = it },
-                    autofill = false
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Preview(showBackground = true)
-@Composable
-fun PreviewWireDialogWith2OptionButtons() {
-    var password by remember { mutableStateOf(TextFieldValue("")) }
-    WireTheme {
-        Box(
-            contentAlignment = Alignment.Center,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            WireDialogContent(
-                optionButton1Properties = WireDialogButtonProperties(
-                    text = "OK",
-                    onClick = { },
-                    type = WireDialogButtonType.Primary,
-                    state = if (password.text.isEmpty()) WireButtonState.Disabled else WireButtonState.Error,
-                ),
-                optionButton2Properties = WireDialogButtonProperties(
-                    text = "Later",
-                    onClick = { },
-                    type = WireDialogButtonType.Primary,
-                    state = if (password.text.isEmpty()) WireButtonState.Disabled else WireButtonState.Error,
-                ),
-                dismissButtonProperties = WireDialogButtonProperties(
-                    text = "Cancel",
-                    onClick = { }
-                ),
-                title = "title",
-                text = buildAnnotatedString {
-                    val style = SpanStyle(
-                        color = colorsScheme().onBackground,
-                        fontWeight = MaterialTheme.wireTypography.body01.fontWeight,
-                        fontSize = MaterialTheme.wireTypography.body01.fontSize,
-                        fontFamily = MaterialTheme.wireTypography.body01.fontFamily,
-                        fontStyle = MaterialTheme.wireTypography.body01.fontStyle
-                    )
-                    withStyle(style) { append("text") }
-                },
-                buttonsHorizontalAlignment = false
-            ) {
-                WirePasswordTextField(
-                    value = password,
-                    onValueChange = { password = it },
-                    autofill = true
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Preview(showBackground = true)
-@Composable
-fun PreviewWireDialogCentered() {
-    var password by remember { mutableStateOf(TextFieldValue("")) }
-    WireTheme {
-        Box(
-            contentAlignment = Alignment.Center,
-            modifier = Modifier.fillMaxWidth()
-        ) {
-            WireDialogContent(
-                optionButton1Properties = WireDialogButtonProperties(
-                    text = "OK",
-                    onClick = { },
-                    type = WireDialogButtonType.Primary,
-                    state = if (password.text.isEmpty()) WireButtonState.Disabled else WireButtonState.Error,
-                ),
-                dismissButtonProperties = WireDialogButtonProperties(
-                    text = "Cancel",
-                    onClick = { }
-                ),
-                centerContent = true,
-                title = "title",
-                text = buildAnnotatedString {
-                    val style = SpanStyle(
-                        color = colorsScheme().onBackground,
-                        fontWeight = MaterialTheme.wireTypography.body01.fontWeight,
-                        fontSize = MaterialTheme.wireTypography.body01.fontSize,
-                        fontFamily = MaterialTheme.wireTypography.body01.fontFamily,
-                        fontStyle = MaterialTheme.wireTypography.body01.fontStyle
-                    )
-                    withStyle(style) { append("text\nsecond line\nthirdLine\nfourth line\nfifth line\nsixth line\nseventh line") }
-                },
-            ) {
-                WirePasswordTextField(
-                    value = password,
-                    onValueChange = { password = it },
-                    autofill = false
-                )
-            }
-        }
-    }
-}
-
 enum class WireDialogButtonType { Primary, Secondary, Tertiary }
 
 data class WireDialogButtonProperties(
@@ -445,3 +299,6 @@ data class WireDialogButtonProperties(
     val type: WireDialogButtonType = WireDialogButtonType.Secondary,
     val loading: Boolean = false
 )
+
+// todo, replace when markdown is moved to commons
+private const val TAG_URL = "linkTag"
